+++ src/Restaurant.js
@@ -1,20 +1,15 @@
-import React, { useState } from "react";
-import RestaurantButton from "./RestaurantButton";
+import React from "react";
+
+import Order from "./Order";
 const Restaurant = () => {
-  // const pizzas = 0;
-  const [orders, setOrders] = useState(0);
-  function orderOne() {
-    setOrders(orders + 1);
-  }
+  
+ 
+
   return (
     <div>
       <h3>Restaurant Orders</h3>
       <ul>
-        <li>
-          Pizzas: {orders}
-          <RestaurantButton handleClick = {orderOne}  /> 
-         
-        </li>
+        <Order />
       </ul>
     </div>
   );
